pyproject.toml
[project]
name = "sd-helper"
version = "0.1.0"
description = "CLI tool for Huawei Cloud Service Delivery Engineers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
sd-helper = "sd_helper.cli:cli"

[tool.uv]
package = true
